# artifacts/api-server/src/routes/export.ts
import { Router, type IRouter } from "express";
import { db } from "@workspace/db";
import {
  userProfiles,
  dietPlans,
  workoutPlans,
  schedules,
  progress,
} from "@workspace/db";
import { eq } from "drizzle-orm";
import type { Achievement, CompletionEvent } from "@workspace/db";
import { computeLevel, xpToNextLevel } from "../lib/gamification";

const router: IRouter = Router();

router.get("/export/:userId", async (req, res) => {
  const { userId } = req.params;
  const format = (req.query["format"] as string) ?? "json";

  const [profile, dietPlan, workoutPlan, schedule, prog] = await Promise.all([
    db.select().from(userProfiles).where(eq(userProfiles.userId, userId)).then((r) => r[0] ?? null),
    db.select().from(dietPlans).where(eq(dietPlans.userId, userId)).then((r) => r[0] ?? null),
    db.select().from(workoutPlans).where(eq(workoutPlans.userId, userId)).then((r) => r[0] ?? null),
    db.select().from(schedules).where(eq(schedules.userId, userId)).then((r) => r[0] ?? null),
    db.select().from(progress).where(eq(progress.userId, userId)).then((r) => r[0] ?? null),
  ]);

  if (!profile) {
    res.status(404).json({ error: `No state found for user '${userId}'` });
    return;
  }

  const xp = prog?.xp ?? 0;
  const level = computeLevel(xp);
  const nextLevel = xpToNextLevel(xp);
  const history = (prog?.history ?? []) as CompletionEvent[];
  const achievements = (prog?.achievements ?? []) as Achievement[];

  const workoutLogs = history.filter((h) => h.type === "workout").length;
  const dietLogs = history.filter((h) => h.type === "diet").length;

  const reportDate = new Date().toISOString().split("T")[0];

  if (format === "json") {
    res.setHeader("Content-Type", "application/json");
    res.setHeader("Content-Disposition", `attachment; filename="fitness-report-${userId}-${reportDate}.json"`);
    res.json({
      exportedAt: new Date().toISOString(),
      userId,
      profile: {
        name: profile.name,
        age: profile.age,
        weightKg: profile.weightKg ? Number(profile.weightKg) : null,
        heightCm: profile.heightCm ? Number(profile.heightCm) : null,
        goal: profile.goal,
        allergies: profile.allergies,
        preferences: profile.preferences,
        availableDays: profile.availableDays,
        equipment: profile.equipment,
        injuries: profile.injuries,
        mode: profile.mode,
      },
      progress: {
        xp,
        streak: prog?.streak ?? 0,
        level,
        xpToNextLevel: nextLevel,
        workoutLogs,
        dietLogs,
        totalLogs: workoutLogs + dietLogs,
      },
      achievements: achievements.map((a) => ({
        name: a.name,
        description: a.description,
        earnedAt: a.earnedAt,
      })),
      dietPlan: dietPlan ? { dailyCalories: dietPlan.dailyCalories, macros: dietPlan.macros, meals: dietPlan.meals } : null,
      workoutPlan: workoutPlan ? { sessions: workoutPlan.sessions } : null,
      schedule: schedule ? { events: schedule.events } : null,
      recentHistory: history.slice(-20),
    });
  } else if (format === "csv") {
    res.setHeader("Content-Type", "text/csv");
    res.setHeader("Content-Disposition", `attachment; filename="fitness-report-${userId}-${reportDate}.csv"`);

    const rows: string[] = [
      "Section,Field,Value",
      `Profile,Name,${profile.name}`,
      `Profile,Goal,${profile.goal}`,
      `Profile,Age,${profile.age ?? ""}`,
      `Profile,Weight (kg),${profile.weightKg ?? ""}`,
      `Profile,Height (cm),${profile.heightCm ?? ""}`,
      `Profile,Available Days,"${(profile.availableDays as string[]).join("; ")}"`,
      `Profile,Equipment,"${(profile.equipment as string[]).join("; ")}"`,
      `Profile,Allergies,"${(profile.allergies as string[]).join("; ")}"`,
      "",
      `Progress,XP,${xp}`,
      `Progress,Level,${level}`,
      `Progress,Streak,${prog?.streak ?? 0}`,
      `Progress,XP to Next Level,${nextLevel}`,
      `Progress,Workout Logs,${workoutLogs}`,
      `Progress,Diet Logs,${dietLogs}`,
      "",
      "Achievements,Name,Earned At",
      ...achievements.map((a) => `Achievements,${a.name},${a.earnedAt.split("T")[0]}`),
      "",
      "History,Type,Completed At,XP Gained,Notes",
      ...history.slice(-20).map((h) => `History,${h.type},${h.completedAt},${h.xpGained},"${h.notes ?? ""}"`),
    ];

    res.send(rows.join("\n"));
  } else if (format === "html") {
    res.setHeader("Content-Type", "text/html");
    res.setHeader("Content-Disposition", `attachment; filename="fitness-report-${userId}-${reportDate}.html"`);

    const achievementRows = achievements
      .map((a) => `<tr><td>${a.name}</td><td>${a.description}</td><td>${a.earnedAt.split("T")[0]}</td><td>+${a.xpBonus} XP</td></tr>`)
      .join("");

    const historyRows = history
      .slice(-20)
      .reverse()
      .map((h) => `<tr><td>${h.type === "workout" ? "💪" : "🥗"} ${h.type}</td><td>${h.completedAt.split("T")[0]}</td><td>+${h.xpGained} XP</td><td>${h.notes ?? ""}</td></tr>`)
      .join("");

    const html = `<!DOCTYPE html>
<html lang="en">
<head>
<meta charset="UTF-8">
<title>Fitness Report — ${profile.name}</title>
<style>
  body { font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif; max-width: 900px; margin: 40px auto; padding: 0 20px; color: #1a1a1a; background: #f9f9f9; }
  h1 { color: #111; border-bottom: 3px solid #0070f3; padding-bottom: 12px; }
  h2 { color: #333; margin-top: 40px; }
  .stats { display: grid; grid-template-columns: repeat(auto-fit, minmax(160px, 1fr)); gap: 16px; margin: 24px 0; }
  .stat { background: white; border-radius: 12px; padding: 20px; box-shadow: 0 2px 8px rgba(0,0,0,0.06); text-align: center; }
  .stat-value { font-size: 2rem; font-weight: 700; color: #0070f3; }
  .stat-label { font-size: 0.85rem; color: #666; margin-top: 4px; }
  table { width: 100%; border-collapse: collapse; background: white; border-radius: 12px; overflow: hidden; box-shadow: 0 2px 8px rgba(0,0,0,0.06); }
  th { background: #0070f3; color: white; padding: 12px 16px; text-align: left; font-weight: 600; }
  td { padding: 10px 16px; border-bottom: 1px solid #f0f0f0; }
  tr:last-child td { border-bottom: none; }
  .badge { background: #fff3cd; color: #856404; padding: 4px 10px; border-radius: 20px; font-size: 0.8rem; font-weight: 600; }
  footer { margin-top: 60px; color: #999; font-size: 0.8rem; text-align: center; }
</style>
</head>
<body>
<h1>🏋️ Fitness Report — ${profile.name}</h1>
<p>Generated on ${reportDate} &nbsp;|&nbsp; Goal: <strong>${profile.goal.replace("_", " ")}</strong> &nbsp;|&nbsp; Mode: <span class="badge">${profile.mode}</span></p>

<h2>Progress</h2>
<div class="stats">
  <div class="stat"><div class="stat-value">${xp.toLocaleString()}</div><div class="stat-label">Total XP</div></div>
  <div class="stat"><div class="stat-value">${level}</div><div class="stat-label">Level</div></div>
  <div class="stat"><div class="stat-value">${prog?.streak ?? 0}</div><div class="stat-label">Day Streak</div></div>
  <div class="stat"><div class="stat-value">${workoutLogs}</div><div class="stat-label">Workouts Logged</div></div>
  <div class="stat"><div class="stat-value">${dietLogs}</div><div class="stat-label">Diet Sessions</div></div>
  <div class="stat"><div class="stat-value">${nextLevel}</div><div class="stat-label">XP to Next Level</div></div>
</div>

${achievements.length > 0 ? `<h2>Achievements (${achievements.length})</h2>
<table>
  <tr><th>Achievement</th><th>Description</th><th>Earned</th><th>Bonus</th></tr>
  ${achievementRows}
</table>` : ""}

<h2>Recent Activity (last 20)</h2>
<table>
  <tr><th>Type</th><th>Date</th><th>XP</th><th>Notes</th></tr>
  ${historyRows || "<tr><td colspan='4'>No activity yet</td></tr>"}
</table>

<footer>Exported from Fitness Agent Layer &nbsp;|&nbsp; ${new Date().toISOString()}</footer>
</body>
</html>`;

    res.send(html);
  } else {
    res.status(400).json({ error: "format must be json, csv, or html" });
  }
});

export default router;