# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: |
    Fitness Agent Layer API — a single backend exposing fitness tools to both
    ChatGPT (Custom GPT Actions) and Claude (MCP server).
    
    Use userId to scope all state to a specific user. A simple string like
    "user_123" or an email is fine for single-user setups.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schedule-events:
    post:
      operationId: scheduleEvents
      tags: [tools]
      summary: Generate and save calendar events for the user's fitness plan
      description: |
        Uses AI to expand a description (e.g. "schedule workouts for a month") into structured calendar events
        based on the user's workout plan and available days.
        In confirm mode, returns a preview without saving — call again with confirmed:true to save.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleEventsRequest"
      responses:
        "200":
          description: Generated events (saved or preview depending on mode)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleEventsResponse"
    delete:
      operationId: clearSchedule
      tags: [tools]
      summary: Clear all scheduled events for a user
      parameters:
        - name: userId
          in: path
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Schedule cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /export/{userId}:
    get:
      operationId: exportReport
      tags: [tools]
      summary: Export a full fitness progress report
      description: |
        Returns a formatted report with profile, progress, achievements, history, diet plan, and workout plan.
        Supports json (default), csv, and html formats.
        The html format is suitable for sharing or printing.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, csv, html]
            default: json
      responses:
        "200":
          description: Report in requested format
          content:
            application/json:
              schema:
                type: object
            text/csv:
              schema:
                type: string
            text/html:
              schema:
                type: string
  /progress/{userId}/reminders/read:
    patch:
      operationId: markRemindersRead
      tags: [tools]
      summary: Mark reminders as read
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
                  description: Reminder IDs to mark as read. Omit to mark all as read.
      responses:
        "200":
          description: Reminders marked as read
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
...
    ScheduleEvent:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        type:
          type: string
          enum: [workout, meal, check_in]
        durationMin:
          type: integer
      required: [title, date, type]
...
    ScheduleEventsRequest:
      type: object
      properties:
        userId:
          type: string
        description:
          type: string
          description: Natural language description (e.g. "schedule workouts for next month")
        startDate:
          type: string
          format: date
          description: Start date (defaults to today)
        durationDays:
          type: integer
          description: Number of days to schedule (default 30)
        confirmed:
          type: boolean
          description: Set to true to save when user is in confirm mode
        eventType:
          type: string
          enum: [workout, meal, both]
      required: [userId]
    ScheduleEventsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleEvent"
        count:
          type: integer
        saved:
          type: boolean
        startDate:
          type: string
        durationDays:
          type: integer
        requiresConfirmation:
          type: boolean
          nullable: true
        message:
          type: string
          nullable: true
      required: [events, count, saved, startDate, durationDays]